Start investigating firmware upgrade with WinBox
I started stepping through the OmniTIK config instructions to see
if we still need the warning not to use the WinBox method.

I didn't finish updating, but cleaned up some of the language and
got some insight into the error and added a note about it.

Also, remove trailing whitespace from lines.