Update realignment guide
I stepped through the draft of the guide using an actual LiteBeam
and fixed the language to reflect the actual menu items and setting
labels. I also referenced the menus and settings in the UISP
mobile app.

I moved the Markdown file to a subdirectory in anticipation of
adding screenshots.